// packages/deltaker-flate-common/model/ulestHendelse.ts
export enum UlestHendelseType {
  InnbyggerGodkjennUtkast = 'InnbyggerGodkjennUtkast',
  NavGodkjennUtkast = 'NavGodkjennUtkast',
  LeggTilOppstartsdato = 'LeggTilOppstartsdato',
  FjernOppstartsdato = 'FjernOppstartsdato',
  EndreStartdato = 'EndreStartdato',
  IkkeAktuell = 'IkkeAktuell',
  AvsluttDeltakelse = 'AvsluttDeltakelse',
  AvbrytDeltakelse = 'AvbrytDeltakelse',
  ReaktiverDeltakelse = 'ReaktiverDeltakelse'
}